requirements.txt
discord.py==0.16.10
aiohttp==2.2.5
beautifulsoup4==4.6.0
psycopg2==2.7.3.1